Task 7 - Simplify Javascript
1 hour

Originally wrote poplateContentArea in spongetest.js to take in three parameters - contentId, templateId and dataKey

When migrating functionality to content JS I decided to use the dataKey as the primary key and updated ids in index.html to facilitate this. Could just as easily pass a map into the content constructor function to dictate what gets rendered where.